feat: CoreBFS
This commit implements CoreBFS, which is enabled by the `-C` command-
line option. To this end, the following additional files and directories
are created:

- external/
  - indexed_bfs/: Contains CoreBFS-related sources
    - generator/: Sources from the Graph500 reference implementation
- mpi/
  - corebfs_adaptor.hpp: Adapter to call code in `external/indexed_bfs`

This configuration is set up to reuse code from Arai's experimental
implementation.

The functionality of CoreBFS itself is added within `BfsBase`. Originally,
I intended to add it in `main.cc`, but due to the wide-ranging impact of
CoreBFS, including `has_edge()` during search key generation and
parameter tuning through `run_bfs()`, it was decided to add it within
`BfsBase` for better organization.

One concern to note is the modification of `BfsValidation::validate()`.
In the current implementation, it appears that validation is performed
using depth information saved during BFS. However, when using CoreBFS,
depth information for the tree portion cannot be obtained.  While the
unused function `build_bfs_depth_map()` provides the depths, it is not
well tested.