fix: remove madvise(2) since it fails on Fugaku
f866357 and f2a42cc introduced `madvise` with `MADV_DONTNEED` for
filling pages with zeros without touching (demanding) those pages.
They are considered to have an effect to reduce the peak memory
consumption. However, it has turned out that `madvise` fails (returns
non-zero) on Fugaku, and allocated pages are not zero-filled. Hence,
this commit reverts pieces of code around `madvise` to those before
f866357 and f2a42cc.

Be careful that this commit may increases the peak memory consumption.
It is needed to investigate the amount of consumed memory before
large-scale execution.